One connection by request (because of socket problems) + struct for matching internal request and replies + cleaning code

+++ src/coap/qcoapmessage.cpp
@@ -8,7 +8,6 @@ QCoapMessagePrivate::QCoapMessagePrivate() :
     type(QCoapMessage::NONCONFIRMABLE),
     messageId(0),
     token(QByteArray()),
-    //tokenLength(0),
     payload(QByteArray())
 {
 }
@@ -19,7 +18,6 @@ QCoapMessagePrivate::QCoapMessagePrivate(const QCoapMessagePrivate& other) :
     type(other.type),
     messageId(other.messageId),
     token(other.token),
-    //tokenLength(other.tokenLength),
     options(other.options),
     payload(other.payload)
 {
@@ -92,7 +90,7 @@ QByteArray QCoapMessage::token() const
 
 quint8 QCoapMessage::tokenLength() const
 {
-    return d_ptr->token.length(); //tokenLength;
+    return static_cast<quint8>(d_ptr->token.length()); //tokenLength;
 }
 
 quint16 QCoapMessage::messageId() const
@@ -138,7 +136,6 @@ void QCoapMessage::setToken(const QByteArray& token)
         return;
 
     d_ptr->token = token;
-    //d_ptr->tokenLength = quint8(token.size());
 }
 
 void QCoapMessage::setMessageId(quint16 id)
